Fix file location problems after moving form validator
